tools/requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0  # Opcional, para comparações visuais